# Trim duplicates