**What this PR does / why we need it**:

Better performance:

```
name                                   old time/op    new time/op    delta
TrimDuplicates/0x0/no-duplicates-36       177ns ± 3%       2ns ± 0%   -98.76%  (p=0.008 n=5+5)
...
TrimDuplicates/9x9/end-duplicate-36      14.1µs ± 1%     3.0µs ± 1%   -78.39%  (p=0.008 n=5+5)

name                                   old alloc/op   new alloc/op   delta
TrimDuplicates/0x0/no-duplicates-36       48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
...
TrimDuplicates/9x9/no-duplicates-36      4.35kB ± 0%    0.90kB ± 0%   -79.42%  (p=0.008 n=5+5)
TrimDuplicates/9x9/all-duplicates-36     2.03kB ± 0%    0.90kB ± 0%   -55.91%  (p=0.008 n=5+5)
TrimDuplicates/9x9/start-duplicate-36    4.71kB ± 0%    0.96kB ± 0%   -79.56%  (p=0.008 n=5+5)
TrimDuplicates/9x9/end-duplicate-36      4.71kB ± 0%    0.96kB ± 0%      ~     (p=0.079 n=4+5)

name                                   old allocs/op  new allocs/op  delta
TrimDuplicates/0x0/no-duplicates-36        1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
...
TrimDuplicates/9x9/end-duplicate-36        39.0 ± 0%       2.0 ± 0%   -94.87%  (p=0.008 n=5+5)
```


**Which issue(s) this PR fixes**:
Related-to #286

**Special notes for your reviewer**:

The new implementation is simpler and handles some cases differently than before. See commit messages for details.

#286 does not get closed by this only because there is some work left for static code analysis.

**Release note**:
```release-note
Handling of key/value pairs is faster. To achieve this, duplicate keys only get removed when they occur in the WithValues key/value pairs and the log call parameters, but not when the parameters of a call like WithValues of InfoS already contain duplicates. That needs to be prevented during code review. Static code analysis with a future release of logcheck will calso catch this.
```