@pohly 2022-05-10
Full benchmark comparison:

<details>

```
name                                   old time/op    new time/op    delta
TrimDuplicates/0x0/no-duplicates-36       177ns ± 3%       2ns ± 0%   -98.76%  (p=0.008 n=5+5)
TrimDuplicates/0x0/all-duplicates-36      179ns ± 5%       2ns ± 1%   -98.77%  (p=0.008 n=5+5)
TrimDuplicates/0x0/start-duplicate-36     408ns ± 3%     305ns ± 3%   -25.23%  (p=0.008 n=5+5)
TrimDuplicates/0x0/end-duplicate-36       401ns ± 3%     302ns ± 7%   -24.70%  (p=0.008 n=5+5)
TrimDuplicates/0x1/no-duplicates-36       377ns ± 2%       3ns ± 2%   -99.29%  (p=0.008 n=5+5)
TrimDuplicates/0x1/all-duplicates-36      382ns ± 5%       3ns ± 1%   -99.29%  (p=0.008 n=5+5)
TrimDuplicates/0x1/start-duplicate-36     844ns ± 4%     396ns ± 4%   -53.15%  (p=0.008 n=5+5)
TrimDuplicates/0x1/end-duplicate-36       833ns ± 7%     405ns ± 4%   -51.36%  (p=0.008 n=5+5)
TrimDuplicates/0x2/no-duplicates-36       778ns ± 1%       3ns ± 2%   -99.66%  (p=0.008 n=5+5)
TrimDuplicates/0x2/all-duplicates-36      764ns ± 7%       3ns ± 1%   -99.65%  (p=0.008 n=5+5)
TrimDuplicates/0x2/start-duplicate-36    1.31µs ± 3%    0.47µs ± 3%   -64.17%  (p=0.008 n=5+5)
TrimDuplicates/0x2/end-duplicate-36      1.30µs ± 5%    0.48µs ± 4%   -63.47%  (p=0.008 n=5+5)
TrimDuplicates/0x3/no-duplicates-36      1.26µs ± 3%    0.00µs ± 1%   -99.79%  (p=0.008 n=5+5)
TrimDuplicates/0x3/all-duplicates-36     1.24µs ± 3%    0.00µs ± 1%   -99.78%  (p=0.008 n=5+5)
TrimDuplicates/0x3/start-duplicate-36    1.85µs ± 3%    0.56µs ± 4%   -69.69%  (p=0.008 n=5+5)
TrimDuplicates/0x3/end-duplicate-36      1.83µs ± 6%    0.57µs ± 3%   -68.69%  (p=0.008 n=5+5)
TrimDuplicates/0x4/no-duplicates-36      1.83µs ± 3%    0.00µs ± 0%   -99.85%  (p=0.016 n=5+4)
TrimDuplicates/0x4/all-duplicates-36     1.79µs ± 1%    0.00µs ± 0%   -99.85%  (p=0.008 n=5+5)
TrimDuplicates/0x4/start-duplicate-36    2.42µs ± 3%    0.66µs ± 4%   -72.65%  (p=0.008 n=5+5)
TrimDuplicates/0x4/end-duplicate-36      2.38µs ± 3%    0.67µs ± 3%   -71.70%  (p=0.008 n=5+5)
TrimDuplicates/0x5/no-duplicates-36      2.36µs ± 2%    0.00µs ± 1%   -99.89%  (p=0.008 n=5+5)
TrimDuplicates/0x5/all-duplicates-36     2.34µs ± 3%    0.00µs ± 1%   -99.89%  (p=0.008 n=5+5)
TrimDuplicates/0x5/start-duplicate-36    3.03µs ± 1%    0.75µs ± 1%   -75.05%  (p=0.008 n=5+5)
TrimDuplicates/0x5/end-duplicate-36      3.05µs ± 2%    0.76µs ± 3%   -74.99%  (p=0.008 n=5+5)
TrimDuplicates/0x6/no-duplicates-36      2.98µs ± 2%    0.00µs ± 1%   -99.91%  (p=0.008 n=5+5)
TrimDuplicates/0x6/all-duplicates-36     2.95µs ± 3%    0.00µs ± 1%   -99.91%  (p=0.008 n=5+5)
TrimDuplicates/0x6/start-duplicate-36    3.68µs ± 3%    0.83µs ± 4%   -77.38%  (p=0.008 n=5+5)
TrimDuplicates/0x6/end-duplicate-36      3.65µs ± 2%    0.85µs ± 3%   -76.66%  (p=0.008 n=5+5)
TrimDuplicates/0x7/no-duplicates-36      3.53µs ± 4%    0.00µs ± 1%   -99.92%  (p=0.008 n=5+5)
TrimDuplicates/0x7/all-duplicates-36     3.57µs ± 3%    0.00µs ± 2%   -99.93%  (p=0.008 n=5+5)
TrimDuplicates/0x7/start-duplicate-36    4.33µs ± 2%    0.92µs ± 3%   -78.66%  (p=0.008 n=5+5)
TrimDuplicates/0x7/end-duplicate-36      4.29µs ± 1%    0.93µs ± 2%   -78.25%  (p=0.008 n=5+5)
TrimDuplicates/0x8/no-duplicates-36      4.21µs ± 1%    0.00µs ± 1%   -99.94%  (p=0.008 n=5+5)
TrimDuplicates/0x8/all-duplicates-36     4.19µs ± 2%    0.00µs ± 0%   -99.94%  (p=0.008 n=5+5)
TrimDuplicates/0x8/start-duplicate-36    5.99µs ± 3%    2.04µs ± 3%   -66.02%  (p=0.008 n=5+5)
TrimDuplicates/0x8/end-duplicate-36      5.93µs ± 2%    2.07µs ± 2%   -65.10%  (p=0.008 n=5+5)
TrimDuplicates/0x9/no-duplicates-36      5.95µs ± 1%    0.00µs ± 1%   -99.95%  (p=0.008 n=5+5)
TrimDuplicates/0x9/all-duplicates-36     5.94µs ± 3%    0.00µs ± 2%   -99.95%  (p=0.008 n=5+5)
TrimDuplicates/0x9/start-duplicate-36    6.78µs ± 1%    2.15µs ± 2%   -68.32%  (p=0.008 n=5+5)
TrimDuplicates/0x9/end-duplicate-36      6.71µs ± 2%    2.19µs ± 1%   -67.33%  (p=0.008 n=5+5)
TrimDuplicates/1x0/no-duplicates-36       386ns ± 1%     161ns ±10%   -58.35%  (p=0.008 n=5+5)
TrimDuplicates/1x0/all-duplicates-36      386ns ± 5%     165ns ± 5%   -57.26%  (p=0.008 n=5+5)
TrimDuplicates/1x0/start-duplicate-36     570ns ± 3%     391ns ± 4%   -31.41%  (p=0.008 n=5+5)
TrimDuplicates/1x0/end-duplicate-36       579ns ± 5%     387ns ± 5%   -33.23%  (p=0.008 n=5+5)
TrimDuplicates/1x1/no-duplicates-36       561ns ± 2%     289ns ± 4%   -48.46%  (p=0.008 n=5+5)
TrimDuplicates/1x1/all-duplicates-36      408ns ± 2%     310ns ± 5%   -24.10%  (p=0.008 n=5+5)
TrimDuplicates/1x1/start-duplicate-36    1.06µs ± 5%    0.49µs ± 2%   -54.41%  (p=0.008 n=5+5)
TrimDuplicates/1x1/end-duplicate-36      1.08µs ± 3%    0.48µs ± 4%   -55.61%  (p=0.008 n=5+5)
TrimDuplicates/1x2/no-duplicates-36       984ns ± 3%     397ns ± 3%   -59.68%  (p=0.008 n=5+5)
TrimDuplicates/1x2/all-duplicates-36      857ns ± 3%     400ns ± 3%   -53.31%  (p=0.008 n=5+5)
TrimDuplicates/1x2/start-duplicate-36    1.51µs ± 2%    0.58µs ± 1%   -61.60%  (p=0.008 n=5+5)
TrimDuplicates/1x2/end-duplicate-36      1.52µs ± 3%    0.58µs ± 1%   -62.07%  (p=0.008 n=5+5)
TrimDuplicates/1x3/no-duplicates-36      1.49µs ± 3%    0.49µs ± 3%   -67.34%  (p=0.008 n=5+5)
TrimDuplicates/1x3/all-duplicates-36     1.32µs ± 1%    0.50µs ± 1%   -62.36%  (p=0.008 n=5+5)
TrimDuplicates/1x3/start-duplicate-36    2.07µs ± 2%    0.66µs ± 1%   -68.40%  (p=0.008 n=5+5)
TrimDuplicates/1x3/end-duplicate-36      2.07µs ± 3%    0.67µs ± 4%   -67.61%  (p=0.008 n=5+5)
TrimDuplicates/1x4/no-duplicates-36      2.02µs ± 2%    0.57µs ± 1%   -71.60%  (p=0.008 n=5+5)
TrimDuplicates/1x4/all-duplicates-36     1.86µs ± 3%    0.59µs ± 4%   -68.33%  (p=0.008 n=5+5)
TrimDuplicates/1x4/start-duplicate-36    2.61µs ± 5%    0.75µs ± 2%   -71.13%  (p=0.008 n=5+5)
TrimDuplicates/1x4/end-duplicate-36      2.66µs ± 3%    0.74µs ± 0%   -72.18%  (p=0.008 n=5+5)
TrimDuplicates/1x5/no-duplicates-36      2.59µs ± 2%    0.66µs ± 4%   -74.63%  (p=0.008 n=5+5)
TrimDuplicates/1x5/all-duplicates-36     2.40µs ± 4%    0.66µs ± 3%   -72.42%  (p=0.008 n=5+5)
TrimDuplicates/1x5/start-duplicate-36    3.22µs ± 2%    0.83µs ± 1%   -74.33%  (p=0.008 n=5+5)
TrimDuplicates/1x5/end-duplicate-36      3.27µs ± 2%    0.83µs ± 3%   -74.44%  (p=0.008 n=5+5)
TrimDuplicates/1x6/no-duplicates-36      3.19µs ± 4%    0.74µs ± 5%   -76.74%  (p=0.008 n=5+5)
TrimDuplicates/1x6/all-duplicates-36     3.04µs ± 2%    0.75µs ± 2%   -75.21%  (p=0.008 n=5+5)
TrimDuplicates/1x6/start-duplicate-36    3.92µs ± 1%    0.93µs ± 2%   -76.27%  (p=0.008 n=5+5)
TrimDuplicates/1x6/end-duplicate-36      3.88µs ± 2%    0.93µs ± 2%   -75.99%  (p=0.008 n=5+5)
TrimDuplicates/1x7/no-duplicates-36      3.87µs ± 1%    0.84µs ± 1%   -78.19%  (p=0.008 n=5+5)
TrimDuplicates/1x7/all-duplicates-36     3.61µs ± 4%    0.84µs ± 3%   -76.64%  (p=0.008 n=5+5)
TrimDuplicates/1x7/start-duplicate-36    5.56µs ± 2%    1.05µs ± 4%   -81.17%  (p=0.008 n=5+5)
TrimDuplicates/1x7/end-duplicate-36      5.50µs ± 2%    1.09µs ± 1%   -80.26%  (p=0.008 n=5+5)
TrimDuplicates/1x8/no-duplicates-36      5.44µs ± 1%    0.97µs ± 2%   -82.23%  (p=0.008 n=5+5)
TrimDuplicates/1x8/all-duplicates-36     4.32µs ± 4%    0.97µs ± 1%   -77.59%  (p=0.008 n=5+5)
TrimDuplicates/1x8/start-duplicate-36    6.27µs ± 2%    2.19µs ± 4%   -65.03%  (p=0.008 n=5+5)
TrimDuplicates/1x8/end-duplicate-36      6.30µs ± 2%    2.16µs ± 2%   -65.80%  (p=0.008 n=5+5)
TrimDuplicates/1x9/no-duplicates-36      6.23µs ± 2%    2.10µs ± 1%   -66.34%  (p=0.008 n=5+5)
TrimDuplicates/1x9/all-duplicates-36     6.08µs ± 0%    2.06µs ± 2%   -66.05%  (p=0.008 n=5+5)
TrimDuplicates/1x9/start-duplicate-36    7.09µs ± 2%    2.28µs ± 1%   -67.84%  (p=0.008 n=5+5)
TrimDuplicates/1x9/end-duplicate-36      7.01µs ± 1%    2.32µs ± 2%   -66.92%  (p=0.008 n=5+5)
TrimDuplicates/2x0/no-duplicates-36       786ns ± 2%     244ns ± 2%   -68.93%  (p=0.008 n=5+5)
TrimDuplicates/2x0/all-duplicates-36      810ns ± 3%     245ns ± 1%   -69.78%  (p=0.008 n=5+5)
TrimDuplicates/2x0/start-duplicate-36    1.06µs ± 3%    0.46µs ± 4%   -57.12%  (p=0.008 n=5+5)
TrimDuplicates/2x0/end-duplicate-36      1.05µs ± 4%    0.46µs ± 4%   -56.18%  (p=0.008 n=5+5)
TrimDuplicates/2x1/no-duplicates-36      1.02µs ± 2%    0.37µs ± 4%   -63.12%  (p=0.008 n=5+5)
TrimDuplicates/2x1/all-duplicates-36      595ns ± 3%     397ns ± 5%   -33.34%  (p=0.008 n=5+5)
TrimDuplicates/2x1/start-duplicate-36    1.50µs ± 2%    0.56µs ± 3%   -62.64%  (p=0.008 n=5+5)
TrimDuplicates/2x1/end-duplicate-36      1.53µs ± 3%    0.56µs ± 3%   -63.52%  (p=0.008 n=5+5)
TrimDuplicates/2x2/no-duplicates-36      1.44µs ± 5%    0.47µs ± 3%   -67.16%  (p=0.008 n=5+5)
TrimDuplicates/2x2/all-duplicates-36      884ns ± 3%     497ns ± 4%   -43.75%  (p=0.008 n=5+5)
TrimDuplicates/2x2/start-duplicate-36    1.98µs ± 5%    0.65µs ± 1%   -67.22%  (p=0.008 n=5+5)
TrimDuplicates/2x2/end-duplicate-36      1.97µs ± 5%    0.65µs ± 2%   -67.07%  (p=0.008 n=5+5)
TrimDuplicates/2x3/no-duplicates-36      1.91µs ± 3%    0.56µs ± 2%   -70.54%  (p=0.008 n=5+5)
TrimDuplicates/2x3/all-duplicates-36     1.37µs ± 5%    0.59µs ± 3%   -56.91%  (p=0.008 n=5+5)
TrimDuplicates/2x3/start-duplicate-36    2.51µs ± 2%    0.74µs ± 1%   -70.37%  (p=0.008 n=5+5)
TrimDuplicates/2x3/end-duplicate-36      2.48µs ± 2%    0.74µs ± 2%   -70.26%  (p=0.008 n=5+5)
TrimDuplicates/2x4/no-duplicates-36      2.47µs ± 6%    0.64µs ± 2%   -74.02%  (p=0.008 n=5+5)
TrimDuplicates/2x4/all-duplicates-36     1.93µs ± 3%    0.66µs ± 5%   -65.74%  (p=0.008 n=5+5)
TrimDuplicates/2x4/start-duplicate-36    3.20µs ± 2%    0.81µs ± 2%   -74.59%  (p=0.008 n=5+5)
TrimDuplicates/2x4/end-duplicate-36      3.09µs ± 4%    0.83µs ± 2%   -73.19%  (p=0.008 n=5+5)
TrimDuplicates/2x5/no-duplicates-36      3.02µs ± 5%    0.73µs ± 4%   -75.66%  (p=0.008 n=5+5)
TrimDuplicates/2x5/all-duplicates-36     2.50µs ± 2%    0.76µs ± 1%   -69.70%  (p=0.008 n=5+5)
TrimDuplicates/2x5/start-duplicate-36    3.76µs ± 3%    0.93µs ± 2%   -75.41%  (p=0.008 n=5+5)
TrimDuplicates/2x5/end-duplicate-36      3.75µs ± 3%    0.94µs ± 2%   -75.01%  (p=0.008 n=5+5)
TrimDuplicates/2x6/no-duplicates-36      3.67µs ± 2%    0.82µs ± 2%   -77.74%  (p=0.008 n=5+5)
TrimDuplicates/2x6/all-duplicates-36     3.12µs ± 2%    0.85µs ± 3%   -72.89%  (p=0.008 n=5+5)
TrimDuplicates/2x6/start-duplicate-36    5.43µs ± 2%    1.00µs ± 4%   -81.62%  (p=0.008 n=5+5)
TrimDuplicates/2x6/end-duplicate-36      5.42µs ± 2%    1.01µs ± 2%   -81.36%  (p=0.008 n=5+5)
TrimDuplicates/2x7/no-duplicates-36      5.36µs ± 3%    0.92µs ± 3%   -82.86%  (p=0.008 n=5+5)
TrimDuplicates/2x7/all-duplicates-36     3.73µs ± 3%    0.93µs ± 2%   -75.01%  (p=0.008 n=5+5)
TrimDuplicates/2x7/start-duplicate-36    6.08µs ± 1%    1.09µs ± 4%   -82.06%  (p=0.008 n=5+5)
TrimDuplicates/2x7/end-duplicate-36      6.00µs ± 3%    1.11µs ± 1%   -81.44%  (p=0.008 n=5+5)
TrimDuplicates/2x8/no-duplicates-36      5.99µs ± 5%    1.02µs ± 3%   -82.99%  (p=0.008 n=5+5)
TrimDuplicates/2x8/all-duplicates-36     4.41µs ± 2%    1.03µs ± 1%   -76.63%  (p=0.008 n=5+5)
TrimDuplicates/2x8/start-duplicate-36    6.77µs ± 2%    2.24µs ± 2%   -66.93%  (p=0.008 n=5+5)
TrimDuplicates/2x8/end-duplicate-36      6.79µs ± 3%    2.30µs ± 1%   -66.17%  (p=0.008 n=5+5)
TrimDuplicates/2x9/no-duplicates-36      6.71µs ± 2%    2.15µs ± 1%   -68.01%  (p=0.008 n=5+5)
TrimDuplicates/2x9/all-duplicates-36     6.02µs ± 2%    2.15µs ± 2%   -64.33%  (p=0.008 n=5+5)
TrimDuplicates/2x9/start-duplicate-36    7.63µs ± 1%    2.37µs ± 3%   -69.01%  (p=0.008 n=5+5)
TrimDuplicates/2x9/end-duplicate-36      7.60µs ± 1%    2.44µs ± 2%   -67.88%  (p=0.008 n=5+5)
TrimDuplicates/3x0/no-duplicates-36      1.29µs ± 2%    0.30µs ± 2%   -76.42%  (p=0.008 n=5+5)
TrimDuplicates/3x0/all-duplicates-36     1.29µs ± 0%    0.30µs ± 1%   -76.80%  (p=0.016 n=4+5)
TrimDuplicates/3x0/start-duplicate-36    1.53µs ± 4%    0.53µs ± 1%   -65.43%  (p=0.008 n=5+5)
TrimDuplicates/3x0/end-duplicate-36      1.51µs ± 2%    0.53µs ± 1%   -65.20%  (p=0.008 n=5+5)
TrimDuplicates/3x1/no-duplicates-36      1.46µs ± 3%    0.45µs ± 2%   -69.47%  (p=0.008 n=5+5)
TrimDuplicates/3x1/all-duplicates-36     1.06µs ± 2%    0.47µs ± 2%   -55.27%  (p=0.008 n=5+5)
TrimDuplicates/3x1/start-duplicate-36    1.99µs ± 3%    0.63µs ± 2%   -68.62%  (p=0.008 n=5+5)
TrimDuplicates/3x1/end-duplicate-36      1.96µs ± 2%    0.62µs ± 2%   -68.26%  (p=0.008 n=5+5)
TrimDuplicates/3x2/no-duplicates-36      1.87µs ± 7%    0.54µs ± 2%   -71.02%  (p=0.008 n=5+5)
TrimDuplicates/3x2/all-duplicates-36     1.05µs ± 4%    0.58µs ± 2%   -44.61%  (p=0.008 n=5+5)
TrimDuplicates/3x2/start-duplicate-36    2.52µs ± 3%    0.72µs ± 1%   -71.54%  (p=0.008 n=5+5)
TrimDuplicates/3x2/end-duplicate-36      2.45µs ± 4%    0.72µs ± 1%   -70.61%  (p=0.008 n=5+5)
TrimDuplicates/3x3/no-duplicates-36      2.43µs ± 3%    0.62µs ± 3%   -74.27%  (p=0.008 n=5+5)
TrimDuplicates/3x3/all-duplicates-36     1.39µs ± 4%    0.67µs ± 2%   -51.55%  (p=0.008 n=5+5)
TrimDuplicates/3x3/start-duplicate-36    3.02µs ± 1%    0.80µs ± 2%   -73.61%  (p=0.008 n=5+5)
TrimDuplicates/3x3/end-duplicate-36      2.96µs ± 5%    0.81µs ± 2%   -72.59%  (p=0.008 n=5+5)
TrimDuplicates/3x4/no-duplicates-36      2.93µs ± 3%    0.72µs ± 4%   -75.51%  (p=0.008 n=5+5)
TrimDuplicates/3x4/all-duplicates-36     2.01µs ± 3%    0.76µs ± 2%   -62.22%  (p=0.008 n=5+5)
TrimDuplicates/3x4/start-duplicate-36    3.65µs ± 5%    0.90µs ± 1%   -75.21%  (p=0.008 n=5+5)
TrimDuplicates/3x4/end-duplicate-36      3.57µs ± 3%    0.92µs ± 2%   -74.21%  (p=0.008 n=5+5)
TrimDuplicates/3x5/no-duplicates-36      3.55µs ± 4%    0.80µs ± 2%   -77.32%  (p=0.008 n=5+5)
TrimDuplicates/3x5/all-duplicates-36     2.51µs ± 5%    0.85µs ± 2%   -66.02%  (p=0.008 n=5+5)
TrimDuplicates/3x5/start-duplicate-36    5.23µs ± 4%    1.01µs ± 2%   -80.69%  (p=0.008 n=5+5)
TrimDuplicates/3x5/end-duplicate-36      5.28µs ± 6%    1.01µs ± 2%   -80.84%  (p=0.008 n=5+5)
TrimDuplicates/3x6/no-duplicates-36      5.15µs ± 2%    0.91µs ± 1%   -82.34%  (p=0.008 n=5+5)
TrimDuplicates/3x6/all-duplicates-36     3.17µs ± 3%    0.94µs ± 4%   -70.42%  (p=0.008 n=5+5)
TrimDuplicates/3x6/start-duplicate-36    5.90µs ± 2%    1.09µs ± 1%   -81.57%  (p=0.008 n=5+5)
TrimDuplicates/3x6/end-duplicate-36      5.94µs ± 0%    1.11µs ± 3%   -81.32%  (p=0.016 n=4+5)
TrimDuplicates/3x7/no-duplicates-36      5.88µs ± 3%    1.00µs ± 3%   -82.94%  (p=0.008 n=5+5)
TrimDuplicates/3x7/all-duplicates-36     3.80µs ± 1%    1.03µs ± 2%   -72.83%  (p=0.008 n=5+5)
TrimDuplicates/3x7/start-duplicate-36    6.56µs ± 6%    1.20µs ± 2%   -81.74%  (p=0.008 n=5+5)
TrimDuplicates/3x7/end-duplicate-36      6.63µs ± 1%    1.22µs ± 2%   -81.56%  (p=0.008 n=5+5)
TrimDuplicates/3x8/no-duplicates-36      6.53µs ± 2%    1.11µs ± 2%   -82.99%  (p=0.008 n=5+5)
TrimDuplicates/3x8/all-duplicates-36     4.44µs ± 2%    1.11µs ± 3%   -74.96%  (p=0.008 n=5+5)
TrimDuplicates/3x8/start-duplicate-36    7.40µs ± 2%    2.37µs ± 2%   -67.97%  (p=0.008 n=5+5)
TrimDuplicates/3x8/end-duplicate-36      7.39µs ± 5%    2.39µs ± 1%   -67.60%  (p=0.008 n=5+5)
TrimDuplicates/3x9/no-duplicates-36      7.32µs ± 2%    2.27µs ± 2%   -68.97%  (p=0.008 n=5+5)
TrimDuplicates/3x9/all-duplicates-36     6.23µs ± 3%    2.28µs ± 2%   -63.35%  (p=0.008 n=5+5)
TrimDuplicates/3x9/start-duplicate-36    8.17µs ± 4%    2.46µs ± 2%   -69.85%  (p=0.008 n=5+5)
TrimDuplicates/3x9/end-duplicate-36      8.31µs ± 1%    2.52µs ± 2%   -69.63%  (p=0.008 n=5+5)
TrimDuplicates/4x0/no-duplicates-36      1.84µs ± 2%    0.36µs ± 2%   -80.61%  (p=0.008 n=5+5)
TrimDuplicates/4x0/all-duplicates-36     1.81µs ± 2%    0.36µs ± 2%   -80.11%  (p=0.008 n=5+5)
TrimDuplicates/4x0/start-duplicate-36    2.06µs ± 5%    0.60µs ± 1%   -70.97%  (p=0.008 n=5+5)
TrimDuplicates/4x0/end-duplicate-36      2.11µs ± 1%    0.60µs ± 3%   -71.69%  (p=0.008 n=5+5)
TrimDuplicates/4x1/no-duplicates-36      2.04µs ± 3%    0.52µs ± 1%   -74.65%  (p=0.008 n=5+5)
TrimDuplicates/4x1/all-duplicates-36     1.54µs ± 4%    0.54µs ± 2%   -65.01%  (p=0.008 n=5+5)
TrimDuplicates/4x1/start-duplicate-36    2.54µs ± 2%    0.69µs ± 3%   -72.75%  (p=0.008 n=5+5)
TrimDuplicates/4x1/end-duplicate-36      2.56µs ± 3%    0.69µs ± 1%   -72.95%  (p=0.008 n=5+5)
TrimDuplicates/4x2/no-duplicates-36      2.48µs ± 3%    0.61µs ± 2%   -75.56%  (p=0.008 n=5+5)
TrimDuplicates/4x2/all-duplicates-36     1.54µs ± 1%    0.65µs ± 2%   -57.90%  (p=0.008 n=5+5)
TrimDuplicates/4x2/start-duplicate-36    3.04µs ± 6%    0.79µs ± 1%   -74.02%  (p=0.008 n=5+5)
TrimDuplicates/4x2/end-duplicate-36      3.08µs ± 1%    0.84µs ± 3%   -72.82%  (p=0.008 n=5+5)
TrimDuplicates/4x3/no-duplicates-36      3.01µs ± 4%    0.71µs ± 1%   -76.43%  (p=0.008 n=5+5)
TrimDuplicates/4x3/all-duplicates-36     1.66µs ± 5%    0.77µs ± 1%   -53.80%  (p=0.008 n=5+5)
TrimDuplicates/4x3/start-duplicate-36    3.65µs ± 1%    0.88µs ± 3%   -75.99%  (p=0.008 n=5+5)
TrimDuplicates/4x3/end-duplicate-36      3.63µs ± 4%    0.95µs ± 1%   -73.85%  (p=0.008 n=5+5)
TrimDuplicates/4x4/no-duplicates-36      3.59µs ± 3%    0.78µs ± 4%   -78.18%  (p=0.008 n=5+5)
TrimDuplicates/4x4/all-duplicates-36     2.07µs ± 2%    0.85µs ± 2%   -58.69%  (p=0.008 n=5+5)
TrimDuplicates/4x4/start-duplicate-36    5.21µs ± 4%    0.98µs ± 2%   -81.18%  (p=0.008 n=5+5)
TrimDuplicates/4x4/end-duplicate-36      5.35µs ± 2%    1.04µs ± 2%   -80.57%  (p=0.008 n=5+5)
TrimDuplicates/4x5/no-duplicates-36      5.19µs ± 4%    0.89µs ± 2%   -82.76%  (p=0.008 n=5+5)
TrimDuplicates/4x5/all-duplicates-36     2.57µs ± 3%    0.96µs ± 4%   -62.77%  (p=0.008 n=5+5)
TrimDuplicates/4x5/start-duplicate-36    5.91µs ± 3%    1.09µs ± 2%   -81.51%  (p=0.008 n=5+5)
TrimDuplicates/4x5/end-duplicate-36      5.94µs ± 2%    1.13µs ± 2%   -81.03%  (p=0.008 n=5+5)
TrimDuplicates/4x6/no-duplicates-36      5.76µs ± 4%    0.99µs ± 1%   -82.82%  (p=0.008 n=5+5)
TrimDuplicates/4x6/all-duplicates-36     3.23µs ± 3%    1.04µs ± 2%   -67.76%  (p=0.008 n=5+5)
TrimDuplicates/4x6/start-duplicate-36    6.56µs ± 3%    1.17µs ± 1%   -82.16%  (p=0.008 n=5+5)
TrimDuplicates/4x6/end-duplicate-36      6.61µs ± 2%    1.24µs ± 1%   -81.27%  (p=0.008 n=5+5)
TrimDuplicates/4x7/no-duplicates-36      6.45µs ± 2%    1.10µs ± 3%   -82.99%  (p=0.008 n=5+5)
TrimDuplicates/4x7/all-duplicates-36     3.87µs ± 2%    1.13µs ± 1%   -70.76%  (p=0.008 n=5+5)
TrimDuplicates/4x7/start-duplicate-36    7.37µs ± 2%    1.30µs ± 2%   -82.39%  (p=0.008 n=5+5)
TrimDuplicates/4x7/end-duplicate-36      7.44µs ± 2%    1.34µs ± 1%   -81.93%  (p=0.008 n=5+5)
TrimDuplicates/4x8/no-duplicates-36      7.26µs ± 3%    1.20µs ± 2%   -83.47%  (p=0.008 n=5+5)
TrimDuplicates/4x8/all-duplicates-36     4.55µs ± 4%    1.23µs ± 2%   -73.08%  (p=0.008 n=5+5)
TrimDuplicates/4x8/start-duplicate-36    8.08µs ± 1%    2.45µs ± 3%   -69.63%  (p=0.008 n=5+5)
TrimDuplicates/4x8/end-duplicate-36      8.01µs ± 2%    2.48µs ± 3%   -69.03%  (p=0.008 n=5+5)
TrimDuplicates/4x9/no-duplicates-36      8.06µs ± 3%    2.37µs ± 2%   -70.56%  (p=0.008 n=5+5)
TrimDuplicates/4x9/all-duplicates-36     6.38µs ± 2%    2.38µs ± 3%   -62.76%  (p=0.008 n=5+5)
TrimDuplicates/4x9/start-duplicate-36    10.6µs ± 2%     2.6µs ± 2%   -75.82%  (p=0.008 n=5+5)
TrimDuplicates/4x9/end-duplicate-36      10.6µs ± 2%     2.6µs ± 1%   -75.20%  (p=0.008 n=5+5)
TrimDuplicates/5x0/no-duplicates-36      2.41µs ± 2%    0.42µs ± 1%   -82.70%  (p=0.008 n=5+5)
TrimDuplicates/5x0/all-duplicates-36     2.40µs ± 5%    0.41µs ± 2%   -82.68%  (p=0.008 n=5+5)
TrimDuplicates/5x0/start-duplicate-36    2.57µs ± 4%    0.68µs ± 1%   -73.48%  (p=0.008 n=5+5)
TrimDuplicates/5x0/end-duplicate-36      2.62µs ± 1%    0.71µs ± 2%   -73.02%  (p=0.008 n=5+5)
TrimDuplicates/5x1/no-duplicates-36      2.56µs ± 2%    0.58µs ± 1%   -77.41%  (p=0.008 n=5+5)
TrimDuplicates/5x1/all-duplicates-36     2.11µs ± 2%    0.61µs ± 1%   -71.07%  (p=0.008 n=5+5)
TrimDuplicates/5x1/start-duplicate-36    3.13µs ± 2%    0.78µs ± 1%   -75.18%  (p=0.008 n=5+5)
…